+++ src/js/engine/initialize.js
@@ -1,0 +1,5 @@
+LiteGraph.debug = false;
+    
+var graph = new LGraph();
+var canvas = new LGraphCanvas("#graphcanvas", graph);
+canvas.resize(1920,1080);