+++ src/js/nodes/default.js
@@ -1,0 +1,3 @@
+var node_alfred = LiteGraph.createNode("Básico/Alfred");
+node_alfred.pos = [17,53];
+graph.add(node_alfred);