+++ src/js/engine/initialize.js
@@ -3,3 +3,4 @@ LiteGraph.debug = false;
 var graph = new LGraph();
 var canvas = new LGraphCanvas("#graphcanvas", graph);
 canvas.resize(1920,1080);
+graph.start();