+++ src/js/engine/initialize.js
@@ -4,3 +4,5 @@ var graph = new LGraph();
 var canvas = new LGraphCanvas("#graphcanvas", graph);
 canvas.resize(1920,1080);
 graph.start();
+
+var variables_table = {};