+++ src/js/nodes/constant/booleano.js
@@ -8,7 +8,7 @@ function Booleano() {
         this.setValue.bind(this)
     );
     this.widgets_up = true;
-    this.addProperty("valor", "")
+    this.addProperty("valor", false);
 }
 
 Booleano.title = "Booleano";
@@ -16,20 +16,20 @@ Booleano.desc = "Crea un texto";
 
 Booleano.prototype.onExecute = function() {
     this.setOutputData(0, this.properties["valor"]);
-};
+}
 
 Booleano.prototype.getTitle = function() {
     if (this.flags.collapsed) {
         return this.properties["valor"] ? "Verdadero" : "Falso"; }
     return this.title;
-};
+}
 
 Booleano.prototype.onPropertyChanged = function(name, value) {
     this.widget.value = value;
-};
+}
 
 Booleano.prototype.setValue = function(v) {
     this.properties["valor"] = v;
-};
+}
 
-LiteGraph.registerNodeType("Variable/Booleano", Booleano);
+LiteGraph.registerNodeType("Constante/Booleano", Booleano);