+++ src/js/nodes/basic/variable.js
@@ -13,15 +13,15 @@ Variable.desc = "Asigna un nombre a un valor";
 
 Variable.prototype.getTitle = function() {
     if(this.flags.collapsed) {
-        let nombre = this.getInputData(1);
+        var nombre = this.getInputData(1);
         return nombre ? nombre : "< Variable >"; }
     return this.title;
 }
 
 Variable.prototype.onAction = function()
 {
-    let input_nombre = this.getInputData(1);
-    let input_value = this.getInputData(2);
+    var input_nombre = this.getInputData(1);
+    var input_value = this.getInputData(2);
 
     if(input_nombre && input_value !== undefined) {
         variables_table[input_nombre] = input_value; }