+++ src/js/nodes/control/condicional.js
@@ -11,7 +11,7 @@ Condicional.desc = "Crear una intersección del flujo principal del programa";
 
 Condicional.prototype.onAction = function()
 {
-    let condition = this.getInputData(1);
+    var condition = this.getInputData(1);
     if(condition === true) { this.triggerSlot(0, "condicional verdadero"); }
     else if(condition === false) { this.triggerSlot(1, "condicional falso"); }
 }