+++ src/js/nodes/operation/comparar.js
@@ -4,7 +4,7 @@ function Comparar()
     this.addInput("A", undefined);
     this.addInput("B", undefined);
     this.addOutput("", LiteGraph.EVENT);
-    this.addOutput("Resultado", LiteGraph.EVENT);
+    this.addOutput("Resultado", undefined);
 }
 
 Comparar.title = "Comparar";
@@ -12,8 +12,8 @@ Comparar.desc = "Compara dos valores y devuelve verdadero o falso";
 
 Comparar.prototype.onAction = function()
 {
-    let A = this.getInputData(1);
-    let B = this.getInputData(2);
+    var A = this.getInputData(1);
+    var B = this.getInputData(2);
     if(A !== undefined && B !== undefined) {
         this.setOutputData(1, A === B);
         this.triggerSlot(0, "comparar");