+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 ##### June 30th, 2019
 
 ###### Updates
+- [x] Modulación del index.html: engine/initialize.js, nodes/default.js
+- [x] Añadido nodo predeterminado: Alfred
+- [x] Añadida restricción para el nodo Alfred (singleton)
+- [x] Añadido nuevo nodo: Alfred
+
 - [x] Nueva rama creada en el repositorio de GitHub
 - [x] Añadidos nuevos nodos: Básico/Muestra, Variable/Texto y Variable/Número
 - [x] Nodos predeterminados eliminados