+++ CHANGELOG.md
@@ -4,10 +4,16 @@
 ##### June 30th, 2019
 
 ###### Updates
+- [x] Añadido nuevo nodo: Cliente/MongoDB
+- [x] Mejorado el nodo de Mostrar: JSON implementado
+- [x] Añadidas las funcionalidades de Node a Electron
+- [x] Añadido el módulo de MongoJS a las dependencias
+- [x] Borrado código innecesario: Alfred.prototype.onExecute()
+
 - [x] Modulación del index.html: engine/initialize.js, nodes/default.js
 - [x] Añadido nodo predeterminado: Alfred
 - [x] Añadida restricción para el nodo Alfred (singleton)
-- [x] Añadido nuevo nodo: Alfred
+- [x] Añadido nuevo nodo: Básico/Alfred
 
 - [x] Nueva rama creada en el repositorio de GitHub
 - [x] Añadidos nuevos nodos: Básico/Muestra, Variable/Texto y Variable/Número
@@ -18,3 +24,4 @@
 - [x] Nueva "visualización" de cómo llevar a cabo el proyecto
 
 ###### Issues
+- [x] Corregido error en el nodo Básico/Alfred que impedia ejecutar graph