+++ CHANGELOG.md
@@ -4,16 +4,25 @@
 ##### June 30th, 2019
 
 ###### Updates
+- [x] Nodo modificado: Básico/Mostrar muestra valores booleanos
+- [x] Añadido nuevo nodo: Control/Condicional
+- [x] Añadido nuevo nodo: Variable/Booleano
+- [x] Nodo modificado: Cliente/MongoDB se maneja mediante eventos o botón
+- [x] Ahora se manejan los nodos mediante eventos
+- [x] Nodo remodelado: Básico/Alfred
+
 - [x] Añadido nuevo nodo: Cliente/MongoDB
-- [x] Mejorado el nodo de Mostrar: JSON implementado
+- [x] Nodo modificado: Básico/Mostrar muestra valores JSON
 - [x] Añadidas las funcionalidades de Node a Electron
 - [x] Añadido el módulo de MongoJS a las dependencias
 - [x] Borrado código innecesario: Alfred.prototype.onExecute()
 
+
 - [x] Modulación del index.html: engine/initialize.js, nodes/default.js
-- [x] Añadido nodo predeterminado: Alfred
-- [x] Añadida restricción para el nodo Alfred (singleton)
+- [x] Añadido nodo predeterminado: Básico/Alfred
+- [x] Añadida restricción para el nodo Básico/Alfred (singleton)
 - [x] Añadido nuevo nodo: Básico/Alfred
+
 
 - [x] Nueva rama creada en el repositorio de GitHub
 - [x] Añadidos nuevos nodos: Básico/Muestra, Variable/Texto y Variable/Número